Extract order permissions to a separate class
